+++ SunEngine/Controllers/AccountController.cs
@@ -13,10 +13,12 @@ using Microsoft.Extensions.Logging;
 using Microsoft.AspNetCore.Identity;
 using SunEngine.Configuration.Options;
 using SunEngine.DataBase;
-using SunEngine.Filters;
 using SunEngine.Managers;
 using SunEngine.Models;
 using SunEngine.Security;
+using SunEngine.Security.Authorization;
+using SunEngine.Security.Captcha;
+using SunEngine.Security.Filters;
 using SunEngine.Services;
 using SunEngine.Stores;
 using SunEngine.Stores.Models;
