+++ SunEngine/Controllers/CaptchaController.cs
@@ -1,7 +1,9 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using SunEngine.Filters;
 using SunEngine.Managers;
+using SunEngine.Security;
+using SunEngine.Security.Captcha;
+using SunEngine.Security.Filters;
 using SunEngine.Services;
 using SunEngine.Stores;
