+++ SunEngine/Controllers/MaterialsController.cs
@@ -1,12 +1,12 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using SunEngine.Filters;
 using SunEngine.Managers;
 using SunEngine.Models;
 using SunEngine.Models.Materials;
 using SunEngine.Presenters;
 using SunEngine.Security.Authorization;
+using SunEngine.Security.Filters;
 using SunEngine.Stores;
 
 namespace SunEngine.Controllers
