+++ SunEngine/Controllers/MessagesController.cs
@@ -1,11 +1,11 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using SunEngine.Filters;
 using SunEngine.Managers;
 using SunEngine.Models.Materials;
 using SunEngine.Presenters;
 using SunEngine.Security.Authorization;
+using SunEngine.Security.Filters;
 using SunEngine.Stores;
 using IAuthorizationService = SunEngine.Security.Authorization.IAuthorizationService;
