+++ SunEngine/Managers/MyUserManager.cs
@@ -19,19 +19,19 @@ namespace SunEngine.Managers
         public MyUserManager(DataBaseConnection db, IUserStore<User> store, IOptions<IdentityOptions> optionsAccessor, IPasswordHasher<User> passwordHasher, IEnumerable<IUserValidator<User>> userValidators, IEnumerable<IPasswordValidator<User>> passwordValidators, ILookupNormalizer keyNormalizer, IdentityErrorDescriber errors, IServiceProvider services, ILogger<UserManager<User>> logger) : base(store, optionsAccessor, passwordHasher, userValidators, passwordValidators, keyNormalizer, errors, services, logger)
         {
             this.db = db;
-            KeyNormalizer = Normalizer.Singleton;
+            KeyNormalizer = FieldNormalizer.Singleton;
         }    
         
         public virtual Task ChangeEmailAsync(int userId, string email)
         {
             return db.Users.Where(x => x.Id == userId).Set(x => x.Email, email)
-                .Set(x => x.NormalizedEmail, Normalizer.Singleton.Normalize(email))
+                .Set(x => x.NormalizedEmail, FieldNormalizer.Singleton.Normalize(email))
                 .UpdateAsync();
         }
 
         public virtual Task<bool> CheckEmailInDbAsync(string email, int userId)
         {
-            return db.Users.AnyAsync(x => x.NormalizedEmail == Normalizer.Singleton.Normalize(email) && x.Id != userId);
+            return db.Users.AnyAsync(x => x.NormalizedEmail == FieldNormalizer.Singleton.Normalize(email) && x.Id != userId);
         }
 
         public virtual LongSession FindLongSession(LongSession longSession)
@@ -48,7 +48,7 @@ namespace SunEngine.Managers
 
         public virtual Task<bool> IsUserInRoleAsync(int userId, string roleName)
         {
-            var normalizedRoleName = Normalizer.Singleton.Normalize(roleName);
+            var normalizedRoleName = FieldNormalizer.Singleton.Normalize(roleName);
             return db.UserToGroups.AnyAsync(x => x.UserId == userId && x.UserGroup.NormalizedName == normalizedRoleName);
         }
     }
