+++ SunEngine/Security/Authentication/MyJwtHanler.cs
@@ -11,6 +11,7 @@ using Microsoft.IdentityModel.Tokens;
 using SunEngine.Configuration.Options;
 using SunEngine.Managers;
 using SunEngine.Models;
+using SunEngine.Security.Authorization;
 using SunEngine.Stores;
 using SunEngine.Stores.Models;
