+++ SunEngine/Security/Captcha/CaptchaValidationFilter.cs
@@ -4,9 +4,8 @@ using Microsoft.AspNetCore.Mvc.Controllers;
 using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.Extensions.DependencyInjection;
 using SunEngine.Controllers;
-using SunEngine.Services;
 
-namespace SunEngine.Filters
+namespace SunEngine.Security.Captcha
 {
     public class CaptchaValidationFilter : ActionFilterAttribute
     {
