+++ SunEngine/Security/Cryptography/CryptoRandomizer.cs
@@ -2,7 +2,7 @@ using System;
 using System.Security.Cryptography;
 using System.Text;
 
-namespace SunEngine.Utils
+namespace SunEngine.Security.Cryptography
 {
     public static class CryptoRandomizer
     {
