+++ SunEngine/Utils/FieldNormalizer.cs
@@ -2,9 +2,9 @@ using Microsoft.AspNetCore.Identity;
 
 namespace SunEngine.Utils
 {
-    public  class Normalizer : ILookupNormalizer
+    public  class FieldNormalizer : ILookupNormalizer
     {
-        public static readonly Normalizer Singleton = new Normalizer();
+        public static readonly FieldNormalizer Singleton = new FieldNormalizer();
         
         public string Normalize(string key)
         {
